+++ parasol/__init__.py
@@ -8,6 +8,6 @@ __all__ = ['ParametricSoln', 'makeSensitivityPlot', 'makeCarpetPlot', 'make2DPar
     'makeContourPlot', 'make2DPlot', 'minimize', 'maximize']
 
 
-from parasol_main import ParametricSoln
-from Optimize import minimize, maximize
-from Plots import *
+from .parasol_main import ParametricSoln
+from .Optimize import minimize, maximize
+from .Plots import *